+++ test_nan_u8.c
@@ -1,0 +1,1 @@
+float g8 = __builtin_nanf(u8"");