+++ test_nan_u8.cpp
@@ -1,0 +1,1 @@
+constexpr float g8 = __builtin_nanf(u8"");