+++ test_nan_wide2.c
@@ -1,0 +1,1 @@
+char hello = __builtin_nanf(L"");